Refactor searching approach to abide the original prompt
The anthropic guide dictated that they were using web search and web fetch, which are normal search and web crawling, not llm-based search like sonar api. But since I'm using sonar as my go-to for now, and since sonar can provide search results with citations too. I've refactored the tool names to match web fetch and search. And also the search approach to be an API call, not a koog single run agent to have more control over the request.
The web_fetch tool is simulated with a scraping approach using sonar.